moved function of background callback to another page (.py)
